requirements.txt
astroid==2.2.5
autopep8==1.4.3
certifi==2019.3.9
cloudinary==1.15.0
coverage==4.5.3
Django==2.1.7
django-cors-middleware==1.2.0
django-extensions==2.1.6
django-nose==1.4.6
djangorestframework==3.9.2
environ==1.0
gunicorn==19.9.0
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
nose==1.3.7
pbr==5.1.3
pep8==1.7.1
Pillow==6.0.0
psycopg2-binary==2.7.7
pycodestyle==2.5.0
PyJWT==1.4.2
pylint==2.3.1
python-decouple==3.1
pytz==2018.9
six==1.10.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
